cloud-fuction/upload_json/requirements.txt
functions-framework==3.*
requests
google-cloud-storage
google-cloud-bigquery